MyApp/requirements.txt
anthropic
pyperclip
pynput
python-dotenv
